Language.Haskell.Names.Annotated

Documentation

data Scoped l

A pair of the name information and original annotation. Used as an annotation type for AST.

Constructors

Scoped (NameInfo l) l

data NameInfo l

Constructors

GlobalValue (SymValueInfo OrigName)
global value

GlobalType (SymTypeInfo OrigName)
global type

LocalValue SrcLoc
local value, and location where it is bound

TypeVar SrcLoc
type variable, and location where it is bound

ValueBinder
here the value name is bound

TypeBinder
here the type name is defined

Import Table
import declaration, and the table of symbols that it introduces

ImportPart Symbols
part of an import declaration

Export Symbols
export declaration, and the symbols it exports

None
no annotation

ScopeError (Error l)
scope error

annotate :: forall a l. (Resolvable (a (Scoped l)), Functor a, Typeable l) => Scope -> a l -> a (Scoped l)