Language.Haskell.Names.GlobalSymbolTable

Description

This module is designed to be imported qualified.

Documentation

data Table

Global symbol table — contains global names

data GName

Possibly qualified name. If the name is not qualified, ModuleNameS is the empty string.

data OrigName

Qualified name, where ModuleNameS points to the module where the name was originally defined. The module part is never empty.

Also contains name and version of the package where it was defined. If it's Nothing, then the entity is defined in the "current" package.

empty :: Table

data Result l a

Constructors

Result a

Error (Error l)

Special

lookupValue :: QName l -> Table -> Result l (SymValueInfo OrigName)

addValue :: QName l -> SymValueInfo OrigName -> Table -> Table

lookupType :: QName l -> Table -> Result l (SymTypeInfo OrigName)

addType :: QName l -> SymTypeInfo OrigName -> Table -> Table

fromMaps :: Map GName (Set (SymValueInfo OrigName)) -> Map GName (Set (SymTypeInfo OrigName)) -> Table

fromLists :: ([(GName, SymValueInfo OrigName)], [(GName, SymTypeInfo OrigName)]) -> Table

types :: Table -> Map GName (Set (SymTypeInfo OrigName))

values :: Table -> Map GName (Set (SymValueInfo OrigName))

toSymbols :: Table -> Symbols